feat: measure time guest writes spend blocked on backpressure (#91)
* fix: restore lint on dev

Two violations landed with the ordered-merge drain classifier and fail
golangci-lint on a clean dev checkout, blocking every branch cut from it.
math/rand/v2 needs a seed pair and spells Intn as IntN; the sequence changes
but the test is a differential guard against a second implementation, so it
needs determinism and fragmentation, not one particular sequence.

* feat: measure time guest writes spend blocked on backpressure

Guest write latency on an EKS control-plane volume measures a 6.2ms mean,
which with the 5.9ms flush beside it accounts for the whole of etcd's 12.1ms
commit. A mean cannot say whether that is every write costing 6ms or a fifth
of them blocking for 30ms, and the two have different fixes.

awaitBackpressure is the candidate: it stalls a guest write and drives a
synchronous drain with a 10-50ms backoff once buffered bytes cross the
high-watermark, and backend writes on that run averaged 356ms. Timing it
separates the stall from the write itself.

Only the blocking path is recorded. The fast path is the common case and
counting it would leave a mean dominated by zeros.

+++ viperblock/chunk_fastpath_test.go
@@ -37,7 +37,7 @@ func assertBlocks(t *testing.T, vb *VB, base uint64, count int, fill byte) {
 	blockSize := uint64(vb.BlockSize)
 	want := bytes.Repeat([]byte{fill}, int(blockSize))
 	for i := range count {
-		got, err := vb.ReadAt((base+uint64(i))*blockSize, uint64(blockSize))
+		got, err := vb.ReadAt((base+uint64(i))*blockSize, blockSize)
 		require.NoError(t, err, "block %d", base+uint64(i))
 		assert.Equal(t, want, got, "block %d", base+uint64(i))
 	}
